Move cache migration to a separate file
(Modified by dlitz@dlitz.net to fix typo)
